requirements.txt
vllm >=0.8.5
openai
google-genai
wandb
datasets
transformers
torch >=2.3.0